% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{all_comparisons}
\alias{all_comparisons}
\title{Retrieve estimates of all effects.}
\usage{
all_comparisons(fit, probs = c(0.025, 0.975))
}
\arguments{
\item{fit}{A \linkS4class{bmerFit}.}

\item{probs}{A vector of quantiles to compute for all group means and pairwise comparisons.}
}
\value{
A named list with an entry for each term in the regression.  Each element is itself a list with two components.
The first, \code{estimates}, contains group means or one-standard-deviation increase effects (see 'Details'). The second,
\code{contrasts} contains pairwise comparisons of the distributions in \code{estimates} when applicable.
}
\description{
The \code{all_comparisons} function computes the posterior distribution of each regression term's effect
in the model (i.e. both main effects and interaction terms with the result depending on the term detials).
This function can only be used if \code{set_contr = TRUE} and \code{scale_cont = TRUE} in \code{\link{bmer_control}}.
}
\details{
In all cases, the estimates returned hold the value of all variables not involved in the term at their average values.
For main effects, if the term is a factor, then group means for each factor level are computed
and pairwise comparisons of each of these are also computed.  If the main effect term is continuous,
then the effect of a one-standard-deviation increase in the effect is computed (which is the same as the raw regression output).

Interactions involving only factors are treated in the same way as main effects, with an estimate for each level of
the interaction term and pairwise comparisons for all interaction levels.  For interactions involving more than one
continuous predictor, the effect of a simultaneous one-standard-deviation increase in all of the continuous predictors
involved is computed (which is the same as the raw regression output).  For interactions involving both
factors and continuous predictors, the effect of a one-standard-deviation
increase in the continuous predictors for each factor level is computed, and pairwise comparisons for group differences
in the effect of this one-standard-deviation increase are also computed.

For unordered factors involving NAs, the estimate for observations set to NA is also computed.  This may or may not have
any real meaning depending on the interpretation of the NA value, and so should be interpreted according to the use of
NA in the data.
}

