% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{bmerBuild-class}
\alias{bmerBuild}
\alias{bmerBuild-class}
\title{An S4 class which contains all the information required to fit a \linkS4class{bmerFit}.}
\description{
A \code{bmerBuild} can be summarized using \code{\link{build_summary}}, which is also the default printing method.
}
\section{Slots}{

\describe{
\item{\code{call}}{The function call which created the bmerBuild.}

\item{\code{control}}{The \linkS4class{bmerControl} used.}

\item{\code{family}}{The regression family ("gaussian" or "binomial").}

\item{\code{frame}}{The data.frame containing the fixed effect variables and random grouping factors.}

\item{\code{x}}{The fixed effect model matrix.}

\item{\code{z}}{A named list where each element is the random effects sparse model matrix for a single random grouping factor.}

\item{\code{name}}{The model_name passed to the function call.}

\item{\code{code}}{The commented Stan code.}

\item{\code{stancode}}{The uncommented Stan code.}

\item{\code{stanmod}}{A character string describing the uncommented Stan code. Used to name and identify previously compiled models for future use.}

\item{\code{data}}{A list to be passed to the \code{data} argument of \code{\link[rstan]{sampling}}.}

\item{\code{variables}}{A list containing information about the variables and random effects structure.}

\item{\code{pars}}{A named list containing the names of the paramters and their dimension names.}

\item{\code{messages}}{A character vector of informational messages (not warnings) returned from \code{\link{build_bmer_model}}.}

\item{\code{warnings}}{A character vector of warnings returned from \code{\link{build_bmer_model}}.}
}}
\seealso{
\code{\link{build_summary}},  \code{\link{get_contrasts}},  \code{\link{cat_code}},  \code{\link{par_names}}
}

