% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/all_generic.R
\docType{class}
\name{bmerBuildSummary-class}
\alias{bmerBuildSummary}
\alias{bmerBuildSummary-class}
\alias{show,bmerBuildSummary-method}
\title{An S4 class to hold the summary of a \code{bmerBuild}.}
\usage{
\S4method{show}{bmerBuildSummary}(object)
}
\arguments{
\item{object}{A \code{bmerBuildSummary}.}
}
\description{
For details see \code{\link{build_summary}}.

Show a \code{bmerBuildSummary}.
}

