% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/all_generic.R
\docType{class}
\name{bmerControl-class}
\alias{bmerControl}
\alias{bmerControl-class}
\alias{show,bmerControl-method}
\title{An S4 class to store control parameters for \code{\link{bmer}}}
\usage{
\S4method{show}{bmerControl}(object)
}
\arguments{
\item{object}{A \code{bmerControl}.}
}
\description{
Stores hyperparameters for the priors, determines the amount of function printing,
and gives instructions for how variable scaling, factor contrasts, and NA values are to be handled.

Show a \code{bmerControl}.
}
\details{
It is very important to note that the default weakly-informative priors are set assuming that \code{scale_cont} and
\code{set_contr} are both set to TRUE. These parameters should only be set to FALSE if you have a lot of experience
with Bayesian modeling and can properly adjust the priors.  See the vignette for more details.
}
\section{Slots}{

\describe{
\item{\code{verbose}}{Logical. If TRUE, summaries of the \linkS4class{bmerBuild} and \linkS4class{bmerFit} are printed when
\code{\link{bmer}} is called. If FALSE (the default), only warning messages (if any) are printed.}

\item{\code{scale_cont}}{Logical. If TRUE (the default), continuous variables are scaled (i.e. mean-centered and divided by their
standard deviation) so that they have mean 0 and standard deviation 1 prior to regression. If FALSE (highly not recommended),
continuous variables are left on their original scales.}

\item{\code{set_contr}}{Logical. If TRUE (the default), sum contrasts are set for unordered factors and scaled orthogonal polynomial
contrasts are set for ordered factors prior to regression. If FALSE (highly not recommended), factor contrasts are not checked.}

\item{\code{estimate_scale_beta}}{Character. One of "P", "yes", or "no". If "yes", the scale of the fixed effect t-distribution prior
is estimated with a hyper-prior which is assigned a half-normal distribution with scale \code{sc_beta}. If "no", the prior scale
is not estimated with a hyper-prior, and is instead given a value of \code{scale_beta}.  If "P" (the default), then
\code{estimate_scale_beta} is interpreted as "yes" if there are 10 or more fixed effect coefficients, and as "no" if there are
fewer than 10 fixed effect coefficients.}

\item{\code{sc_beta}}{Numeric (default 1; must be greater than 0). The scale for the half-normal hyper-prior on the scale of the
fixed effect t-distribution prior if a hyper-prior is assigned (depends on \code{estimate_scale_beta}).}

\item{\code{scale_beta}}{Numeric (default 2; must be greater than 0). The scale for the fixed effect t-distribution prior if a
hyper-prior is not assigned (depends on \code{estimate_scale_beta}).}

\item{\code{nu_beta}}{Numeric (default 5; must be greater than 0). The degrees of freedom for the fixed effect t-distribution prior.
Setting to 1 results in a Cauchy prior.}

\item{\code{sc_q0}}{Numeric (default is 1.5; must be greater than 0). The scale for the half-normal prior on the standard deviation in
the random intercepts.}

\item{\code{sc_qs}}{Numeric (default is 1; must be greater than 0). The scale for the half-normal prior on the standard deviation in
the random slopes.}

\item{\code{eta_q}}{Numeric (default is 2; must be greater than 0). The shape parameter for the LKJ prior on the random effects
correlation matrices. Setting to 1 results in a flat prior.}

\item{\code{sc_res}}{Numeric (default is 0.5; must be greater than 0). The scale for the half-normal prior on the standard deviation in
the residuals.  Only applies to linear (i.e. Gaussian) regressions.}

\item{\code{fixef_na}}{Logical. If FALSE (the default), observations with NA values in the fixed effects are dropped. If TRUE, then
\code{set_contr} must also be TRUE, and NA values will be set to 0 in the model matrix (i.e. level slashing). When an unordered
factor with NA values is involved in interactions, the interaction term is checked for higher-order NA values and, if necessary,
the contrasts for the factors with NA values are changed for the interaction term, and these new contrasts which apply to the
interaction term but not the main effects are logged. In regression output, interaction terms with contrast changes are prefixed
with "nai_" so that they can be easily identified. See the vignette for examples.}

\item{\code{ranef_na}}{Logical. If FALSE (the default), observations with NA values in the random effect grouping factors are dropped.
If TRUE, then set_contr must also be TRUE, and NA values will be set to 0 in the model matrix. When a within-group fixed effect
factor only takes on a subset of its values when the random grouping factor is not NA, then the contrasts for the factor are
changed in the coding of slopes for the random effect grouping factor, and the new contrasts which apply to the slope but not the
fixed effect are logged. In the regression output, these slopes are prefixed with the name of the random effect grouping factor
so they can be easily identified.}
}}

