% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{bmerFit-class}
\alias{bmerFit}
\alias{bmerFit-class}
\title{An S4 class to store a \linkS4class{bmerBuild} and \linkS4class{stanfit}.}
\description{
A \code{bmerFit} can be summarized using \code{\link{fit_summary}}, which is also the default printing method.
}
\section{Slots}{

\describe{
\item{\code{call}}{The function call which produced the object.}

\item{\code{build}}{The \linkS4class{bmerBuild} passed to the \code{\link{fit_bmer_build}} function call.}

\item{\code{diagnostics}}{A list of diagnostics from the sampler and for Rhat and n_eff.}

\item{\code{warnings}}{A character vector of warnings returned by \code{\link{fit_bmer_build}} (if any).}
}}
\seealso{
\code{\link{fit_summary}},  \code{\link{get_build}},  \code{\link{named_extract}}, 
\code{\link{all_comparisons}}, \code{\link{bmers_summary}}
}

