% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/all_generic.R
\docType{class}
\name{bmerFitSummary-class}
\alias{bmerFitSummary}
\alias{bmerFitSummary-class}
\alias{show,bmerFitSummary-method}
\title{An S4 class to hold the summary of a \code{bmerFit}.}
\usage{
\S4method{show}{bmerFitSummary}(object)
}
\arguments{
\item{object}{A \code{bmerFitSummary}.}
}
\description{
For details see \code{\link{fit_summary}}.

Show a \code{bmerFitSummary}.
}

