% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{bmers_summary}
\alias{bmers_summary}
\title{Get parameter summary from a \linkS4class{bmerFit}.}
\usage{
bmers_summary(fit, pars = names(par_names(fit)), probs = c(0.025, 0.975),
  interval = c(0, Inf))
}
\arguments{
\item{fit}{A \linkS4class{bmerFit}.}

\item{pars}{Names of the parameters to be summarized (defaults to all parameters)}

\item{probs}{Quantiles to return for each parameter (defaults to a 95\% equal-tailed credible interval)}

\item{interval}{The interval on which the posterior probability of each parameter is to be computed.  The
default is (0, Inf) which returns the posterior probability that each parameter is positive.}
}
\value{
A matrix summarizing the indicated parameters.
}
\description{
Returns a \linkS4class{stanfit} summary along with the posterior probability that a parameter is on a given interval.
}

