% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{build_summary}
\alias{build_summary}
\title{Summarize a \linkS4class{bmerBuild}.}
\usage{
build_summary(object)
}
\arguments{
\item{object}{A \linkS4class{bmerBuild} or \linkS4class{bmerFit}.}
}
\value{
A list of class \linkS4class{bmerBuildSummary} whose elements are given in 'Details'.
}
\description{
Returns a list which contains summary information about a \linkS4class{bmerBuild}. While the individual
elements of the list can be viewed and used, it is most useful for its default show method,
which lays out the important aspects of the model which will be fit when the \linkS4class{bmerBuild}
is passed to \code{\link{fit_bmer_build}}.
}
\details{
The elements of the returned list are:

\strong{call}.  The function call evaluated to create the \linkS4class{bmerBuild}.

\strong{name}.  The name given to the model.

\strong{family}.  The regression family.

\strong{frame}.  The model frame generated from evaluating the formula.  It has attributes "fixed_varmat"
    and "random_varmat" indicating how the variables relate to interaction terms.

\strong{N}.  The number of observations.

\strong{PQ}.  The total number of unconstrained parameters (both fixed and random).

\strong{response}.  A list giving information about the dependent variable: name; class; if binomial,
the levels of the response factor and which was assigned to 0 and to 1; if Gaussian, the raw mean
and standard deviation and a logical indicating whether or not the response is on unit scale in the regression.

\strong{fixed}.  A list giving information about the fixed effects, with three elements: \emph{factors}, a named list
with an entry for each factor in the regression containing its number of levels, contrasts, etc.; \emph{covariates}, a
named list with an entry for each continuous predictor in the regression containing its mean and standard deviation in the
raw data and a logical indicating whether or not it was scaled in the regression; and \emph{terms}, a character vector of
the main effects and interactions in the regression from the expanded fixed effects formula.  If interactions involving
fixed effects with NAs resulted in changing factor contrasts, then additionally there will be a fourth element \emph{nainter},
which is a named list with an element for each interaction term for which contrasts were altered, giving the contrasts
used in the computation of the interaction term.  See the vignette for more details.

\strong{random}.  A named list with an element for each random effects grouping factor, each of which is a list with the
number of levels (i.e. group members), their names, whether or not there are NA values, a list of new factor contrasts
created (if any), a list of slope interactions where factor contrasts were changed (if any) or dropped due to redundancy
(if any), the terms which were evaluated as potential slopes, the terms which were included as slopes, and the names
of the random effects model matrix columns.  See the vignette for more details.

\strong{code}.  The Stan code used to fit the model, along with model-specific comments.  See \code{\link{cat_code}}.

\strong{data}.  The list passed as the \code{data} argument to \code{\link[rstan]{sampling}}.

\strong{control}.  The \code{\link{bmer_control}} used to build the model.

\strong{messages}.  The messages (not warnings) returned from \code{\link{build_bmer_model}} (if any).

\strong{warnings}.  The warnings returned from \code{\link{build_bmer_model}} (if any).
}
\seealso{
\code{\link{build_summary}},  \code{\link{get_contrasts}},  \code{\link{cat_code}},  \code{\link{par_names}},
\code{\link{get_control}}.
}

