% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{cat_code}
\alias{cat_code}
\title{Print or save Stan code.}
\usage{
cat_code(object, file = NULL, commented = TRUE)
}
\arguments{
\item{object}{A \linkS4class{bmerBuild} or \linkS4class{bmerFit}.}

\item{file}{Name of a file to create and write the code to.  If NULL (the default) then the code is printed within R.}

\item{commented}{Logical.  If FALSE, then the code used within the \linkS4class{stanmodel} is used.  If TRUE (the recommended default),
then a version of the code with additional comments specific to the current model is used (but not different in execution from the raw code).}
}
\description{
The \code{cat_code} function extracts the Stan code used in a \code{bmer} model and either
prints it within R or writes it to a file depending on the value of \code{file}.
}

