% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{fit_summary}
\alias{fit_summary}
\title{Summarize a \linkS4class{bmerFit}.}
\usage{
fit_summary(fit)
}
\arguments{
\item{fit}{A \linkS4class{bmerFit}.}
}
\value{
A list of class \linkS4class{bmerFitSummary} whose elements are given in 'Details'.
}
\description{
Returns a list which contains summary information about a bmerFit. The default printing option
gives a summary of the random effects standard deviations and correlations (posterior means), as well
as the fixed effects (posterior means, standard deviations, 95\% credible intervals, effect sample sizes,
Gelman-Rubin Rhat statistics, and posterior probability that each is positive)
}
\details{
The elements of the returned list are:

\strong{call}.  The function call evaluated to create the \linkS4class{bmerFit}.

\strong{build}.  A list returned by \code{\link{build_summary}} summarizing the \linkS4class{bmerBuild} used to fit the \linkS4class{bmerFit}.

\strong{random}.  A named list with an element for each random effect grouping factor, each being a list
with the following elements: \emph{ranef}, the intercept and slope estimates for individual member of the grouping factor;
\emph{coef}, the coefficient vector for each group member obtained from summing the random intercept and slope estimates
with the fixed effects coefficients; \emph{sd}, the standard deviation in the random effects; \emph{cor} the correlation
matrix for the random effects; and \emph{cov} the covariance matrix for the random effects.

\strong{fixed}.  A list with the following elements: \emph{coef}, the fixed effect estiamtes, including the posterior
probability that each is greater than zero (see the vignette for more details); \emph{cor} the correlation in
the fixed effects; and \emph{cov}, the covariance matrix for the fixed effects.

\strong{fitted.values}.  A vector containing the mean estimate for each observation (predicted response value for Gaussian
regression and predicted log-odds of being in category 1 for binomial regression).

\strong{warnings}.  The warnings returned by \code{\link{fit_bmer_build}} (if any).
}
\seealso{
\code{\link{fit_summary}},  \code{\link{get_build}},  \code{\link{named_extract}}, 
\code{\link{get_contrasts}},  \code{\link{cat_code}}, \code{\link{par_names}},  \code{\link{all_comparisons}},
\code{\link{bmers_summary}},  \code{\link{get_control}}.
}

