% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_build}
\alias{get_build}
\title{Get the \linkS4class{bmerBuild} stored in a \linkS4class{bmerFit}.}
\usage{
get_build(object)
}
\arguments{
\item{object}{A \linkS4class{bmerFit} or \linkS4class{bmerBuild}.}
}
\description{
If the function is called on a \linkS4class{bmerFit} then the \linkS4class{bmerBuild}
used to fit the model is returned. If the function is called on a \linkS4class{bmerBuild},
then the same build is returned.
}

