% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Get factor contrasts.}
\usage{
get_contrasts(object)
}
\arguments{
\item{object}{A \linkS4class{bmerBuild} or \linkS4class{bmerFit}.}
}
\value{
A list of contrast matrices organized in named sublists.
}
\description{
\code{get_contrasts} lists the contrast matrix for each factor in the regression.
}
\details{
If there are no factors in the regression, then an empty list is returned.
Otherwise, the list will have an element \code{main_effects} which is itself
a named list with an entry for the contrast matrix for each factor in the regression.
If \code{fixef_na = TRUE} in \code{\link{bmer_control}} and there were interactions
involving unordered factors with NA values which required a change in contrasts for
the interaction term, then in addition to \code{main_effects} there will be an element
\code{na_interactions} with an entry for each interaction where contrasts were changed,
with these entries having the same structure as \code{main_effects}.  Similarly,
if \code{ranef_na = TRUE} in \code{\link{bmer_control}} and there are fixed effects
with levels which only occur when the random effect group is NA, then these factors
will have entries structured the same as \code{main_effects} in a sublist named
with the random effect group followed by \code{_altered_main_effects}.  Interactions
in the slopes for a random effect group which require contrast alterations relative to the way
the main effect random slopes are coded are handled in the same way as fixed effect alterations
relative the fixed main effects are handled, and stored in a sublist named with the random
effect group followed by \code{_na_interactions}.  In the most common use of \code{bmer},
where both \code{fixef_na = FALSE} and \code{ranef_na = FALSE}, the result will only contain
the \code{main_effects} sublist.  For examples of changes due to NAs, see the vignette.
}

