% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_control}
\alias{get_control}
\title{See the control parameters applied to a \linkS4class{bmerBuild} or \linkS4class{bmerFit}.}
\usage{
get_control(object)
}
\arguments{
\item{object}{A \linkS4class{bmerBuild} or \linkS4class{bmerFit}.}
}
\description{
See the control parameters applied to a \linkS4class{bmerBuild} or \linkS4class{bmerFit}.
}

