% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{named_extract}
\alias{named_extract}
\title{Extract posterior samples with named dimensions.}
\usage{
named_extract(fit, pars = names(par_names(fit)))
}
\arguments{
\item{fit}{A \linkS4class{bmerFit}.}

\item{pars}{A character vector of parameter names. See \code{\link{par_names}}.}
}
\value{
A named list where each component contains the posterior samples for a parameter,
and each array of posterior samples has named dimensions.
}
\description{
The function calls \code{\link[rstan]{extract}} and then gives names to the
dimensions of the posterior samples.
}
\details{
As an example, if there were 4 chains
and 1000 post-warmup iterations in each chain, then extracting the random effect
estimates for individual members of a random effects group (the \code{gamma}
parameter for the group) with \code{M} members and \code{Q} effects will return an array
which is \code{4000 x M x Q},
where the names of the second dimension are the group member names and the names
of the third are the random effect names.
}

