% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{par_names}
\alias{par_names}
\title{Get parameter names for a \code{bmer} model.}
\usage{
par_names(object)
}
\arguments{
\item{object}{A \linkS4class{bmerBuild} or \linkS4class{bmerFit}.}
}
\value{
A named list with an entry for each parameter in the model, where the entries have
the names of each dimension in the parameter.
}
\description{
Get parameter names for a \code{bmer} model.
}

