% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\docType{methods}
\name{show,bmerBuild-method}
\alias{show,bmerBuild-method}
\title{Show a \code{\link{build_summary}} for a \linkS4class{bmerBuild}.}
\usage{
\S4method{show}{bmerBuild}(object)
}
\arguments{
\item{object}{A \linkS4class{bmerBuild}.}
}
\description{
Show a \code{\link{build_summary}} for a \linkS4class{bmerBuild}.
}

