% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\docType{methods}
\name{show,bmerFit-method}
\alias{show,bmerFit-method}
\title{Show a \code{\link{fit_summary}} for a \linkS4class{bmerFit}.}
\usage{
\S4method{show}{bmerFit}(object)
}
\arguments{
\item{object}{A \linkS4class{bmerFit}.}
}
\description{
Show a \code{\link{fit_summary}} for a \linkS4class{bmerFit}.
}

